+++ Modules/Intent.Modules.VisualStudio.Projects/release-notes.md
@@ -1,6 +1,6 @@
 ### Version 3.3.30
 
-- Improvement: 
+- Improvement: `launchsettings.json` now listens for an event named `LaunchProfileHttpPortRequired` and if received will add a regular `http://…` entry to a profile's `applicationUrl` when it only contains a single `https://…` item.
 
 ### Version 3.3.29